Fix regression with inlined type with methods
In v2, we report an error if an inlined struct type also has
marshal or unmarshal methods. In such a case, it is ambiguous how
we are supposed to inline the type. The presence of methods
indicates that the type has custom serialization behavior,
but inlining seems to suggest that we should serialize the type
as a Go struct.

Instead of skipping the field, still add the field to the queue
so that we can process the type's fields as if
they were JSON serializable.