Reuse opts in UnmarshalDecode
When writing our own UnmarshalJSONFrom func, we may call back to the
UnmarshalDecode to handle each field, and just pass in the opt unmodified. In
this case, we can just use this opt object and don't need to fetch one from the
pool. This brings over 10% speedup for decoding just a single field.