Support benchmarking v1in2
Currently, we have the ability to benchmark "encoding/json" or
"github.com/go-json-experiment/json" based on an environment variable.
We want to have the ability to also benchmark
"github.com/go-json-experiment/json/v1".

This refactors the benchmark code such that it switches on the version
in a single place rather than spread throughout the code.