Combine `nocase` and `strictcase` tag options under a `case` option
WARNING: This commit contains breaking changes.

Instead of specifying `nocase` or `strictcase` tag options,
use a single `case` tag option, which is a key-value pair
where the value can only be 'ignore' or 'strict'.
This reads more naturally as `case:ignore` or `case:strict` and
is more clear that `case:ignore` and `case:strict` are
mutually exclusive tag options.