jsontext: remove support for write Float(+Inf) as "Infinity"
Similarly, Parsing "Infinity" as +Inf is also removed

This feature is not standard and should not be enabled by default.  If this is
really desired, user can still write a wrapper that invokes Float or String
conditionally. Instead of embedding the maybe undesired branches in jsontext.

It is harder to get consistent API design with the new RawToken refactor. Now
Token.Float() should only process tokens returned by jsontext.Float, we need to
differentate infinity float and infinity string. So what kind should the
infinity float be?
- if '0': Good for that it cannot be used as object key, but can make
  Token.String() confusing; also confusing for that a '0' token actually
  encodes to string.
- if '"': Need many extra branching in Token.Kind and Token.appendString. And
  the kind returned by jsontext.Float() is not consistent.

So, just remove these.