Drop Options argument from user-declared methods and functions
WARNING: This commit contains breaking changes.

This drops Options as an argument from
the MarshalerTo and UnmarshalerFrom interfaces and
the MarshalToFunc and UnmarshalFromFunc functions.

Instead, the options is stored within the
jsontext.Encoder or jsontext.Decoder and
can be retrieved through the Options method.

This simplifies the API for custom marshalers and unmarshalers and
makes it impossible to accidentally drop the options
when recursively calling json.MarshalEncode or json.UnmarshalDecode
from within a custom marshaler or unmarshaler implementation.

Fixes golang/go#71611