jsontext: proof-of-concept ReadStringAsStream
This is an idea for making it possible to have explicit control over
what happens when reading strings. For example, an implementation
could choose to decode on the fly or discard large strings but still
keep going without using a large amount of memory.

We include a somewhat involved example demonstrating some possibilities.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>