jsontext: Preserve the internal buffer between Reset calls
The jsontext decoder throws away the internal buffer on each Reset call.
This makes decoding one message at a time pretty wasteful since each new message
will cause at least one allocation for the buffer.

This commit makes sure the buffer gets reused when the decoder is reset with a new reader.
If unbounded growth for the slice is a concern, I can add an option limiting the maxiumum size
after which the bytes buffer will be thrown away and a new one can be allocated.