Update to 6e618cd42a1adb58fa04f7a9f6e89a563ccb07f1
WARNING: This commit includes a breaking change.
Trying to marshal a time.Duration without an explicit format flag
results in an error. Specify `format:units` to preserve prior behavior.
The purpose of this change is to keep the future open for chosing
a default JSON representation for a duration (see https://go.dev/issue/71631).

This pulls in the following changes:
* (https://go.dev/cl/673335) encoding/json: avoid supurious synctest deadlock detection
* (https://go.dev/cl/680715) encoding/json/jsontext, encoding/json/v2: document experimental nature
* (https://go.dev/cl/681795) encoding/json/jsontext: fix spelling error
* (https://go.dev/cl/683155) encoding/json/jsontext: consistently use JSON terminology
* (https://go.dev/cl/682455) encoding/json/v2: report error on time.Duration without explicit format
* (https://go.dev/cl/683175) encoding/json/v2: reject unquoted dash as a JSON field name
* (https://go.dev/cl/682403) encoding/json/v2: support ISO 8601 durations
* (https://go.dev/cl/525516) encoding/json: use zstd compressed testdata

Also, migrate.sh was modified to exclude TestSynctestMarshal since
package "synctest" is not yet available until Go 1.25.